# Code/Servidor/swagger_server/models/account_item.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server.models.friend_item import FriendItem  # noqa: F401,E501
from swagger_server.models.playlist_item import PlaylistItem  # noqa: F401,E501
from swagger_server import util


class AccountItem(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """

    def __init__(self, id: str=None, username: str=None, name: str=None, bio: str=None, mail: str=None, friends: List[FriendItem]=None, playlists: List[PlaylistItem]=None):  # noqa: E501
        """AccountItem - a model defined in Swagger

        :param id: The id of this AccountItem.  # noqa: E501
        :type id: str
        :param username: The username of this AccountItem.  # noqa: E501
        :type username: str
        :param name: The name of this AccountItem.  # noqa: E501
        :type name: str
        :param bio: The bio of this AccountItem.  # noqa: E501
        :type bio: str
        :param mail: The mail of this AccountItem.  # noqa: E501
        :type mail: str
        :param friends: The friends of this AccountItem.  # noqa: E501
        :type friends: List[FriendItem]
        :param playlists: The playlists of this AccountItem.  # noqa: E501
        :type playlists: List[PlaylistItem]
        """
        self.swagger_types = {
            'id': str,
            'username': str,
            'name': str,
            'bio': str,
            'mail': str,
            'friends': List[FriendItem],
            'playlists': List[PlaylistItem]
        }

        self.attribute_map = {
            'id': 'id',
            'username': 'username',
            'name': 'name',
            'bio': 'bio',
            'mail': 'mail',
            'friends': 'friends',
            'playlists': 'playlists'
        }

        self._id = id
        self._username = username
        self._name = name
        self._bio = bio
        self._mail = mail
        self._friends = friends
        self._playlists = playlists

    @classmethod
    def from_dict(cls, dikt) -> 'AccountItem':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The AccountItem of this AccountItem.  # noqa: E501
        :rtype: AccountItem
        """
        return util.deserialize_model(dikt, cls)

    @property
    def id(self) -> str:
        """Gets the id of this AccountItem.


        :return: The id of this AccountItem.
        :rtype: str
        """
        return self._id

    @id.setter
    def id(self, id: str):
        """Sets the id of this AccountItem.


        :param id: The id of this AccountItem.
        :type id: str
        """
        if id is None:
            raise ValueError("Invalid value for `id`, must not be `None`")  # noqa: E501

        self._id = id

    @property
    def username(self) -> str:
        """Gets the username of this AccountItem.


        :return: The username of this AccountItem.
        :rtype: str
        """
        return self._username

    @username.setter
    def username(self, username: str):
        """Sets the username of this AccountItem.


        :param username: The username of this AccountItem.
        :type username: str
        """
        if username is None:
            raise ValueError("Invalid value for `username`, must not be `None`")  # noqa: E501

        self._username = username

    @property
    def name(self) -> str:
        """Gets the name of this AccountItem.


        :return: The name of this AccountItem.
        :rtype: str
        """
        return self._name

    @name.setter
    def name(self, name: str):
        """Sets the name of this AccountItem.


        :param name: The name of this AccountItem.
        :type name: str
        """
        if name is None:
            raise ValueError("Invalid value for `name`, must not be `None`")  # noqa: E501

        self._name = name

    @property
    def bio(self) -> str:
        """Gets the bio of this AccountItem.


        :return: The bio of this AccountItem.
        :rtype: str
        """
        return self._bio

    @bio.setter
    def bio(self, bio: str):
        """Sets the bio of this AccountItem.


        :param bio: The bio of this AccountItem.
        :type bio: str
        """
        if bio is None:
            raise ValueError("Invalid value for `bio`, must not be `None`")  # noqa: E501

        self._bio = bio

    @property
    def mail(self) -> str:
        """Gets the mail of this AccountItem.


        :return: The mail of this AccountItem.
        :rtype: str
        """
        return self._mail

    @mail.setter
    def mail(self, mail: str):
        """Sets the mail of this AccountItem.


        :param mail: The mail of this AccountItem.
        :type mail: str
        """
        if mail is None:
            raise ValueError("Invalid value for `mail`, must not be `None`")  # noqa: E501

        self._mail = mail

    @property
    def friends(self) -> List[FriendItem]:
        """Gets the friends of this AccountItem.


        :return: The friends of this AccountItem.
        :rtype: List[FriendItem]
        """
        return self._friends

    @friends.setter
    def friends(self, friends: List[FriendItem]):
        """Sets the friends of this AccountItem.


        :param friends: The friends of this AccountItem.
        :type friends: List[FriendItem]
        """
        if friends is None:
            raise ValueError("Invalid value for `friends`, must not be `None`")  # noqa: E501

        self._friends = friends

    @property
    def playlists(self) -> List[PlaylistItem]:
        """Gets the playlists of this AccountItem.


        :return: The playlists of this AccountItem.
        :rtype: List[PlaylistItem]
        """
        return self._playlists

    @playlists.setter
    def playlists(self, playlists: List[PlaylistItem]):
        """Sets the playlists of this AccountItem.


        :param playlists: The playlists of this AccountItem.
        :type playlists: List[PlaylistItem]
        """
        if playlists is None:
            raise ValueError("Invalid value for `playlists`, must not be `None`")  # noqa: E501

        self._playlists = playlists