# Code/Servidor/swagger_server/swagger/swagger.yaml
...
swagger: "2.0"
info:
  description: "Back-end API del Proyecto Cierzo"
  version: "2.0.0"
  title: "Proyecto Cierzo API"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "localhost:8080"
basePath: "/api/"
...
schemes:
- "https"
paths:
...
  /playlists:
    get:
      tags:
      - "public"
      summary: "busca listas de reproducción con ciertos parámetros"
      description: "Al pasarle ciertos parámetros devuelve listas de reproducción\
        \ que se ajusten a ellos\n"
      operationId: "search_playlist"
      produces:
      - "application/json"
      parameters:
      - name: "name"
        in: "query"
        description: "nombre de la lista"
        required: false
        type: "string"
      - name: "owner"
        in: "query"
        description: "propietario de la lista"
        required: false
        type: "string"
      - name: "skip"
        in: "query"
        description: "number of records to skip for pagination"
        required: false
        type: "integer"
        minimum: 0
        format: "int32"
      - name: "limit"
        in: "query"
        description: "maximum number of records to return"
        required: false
        type: "integer"
        maximum: 50.0
        minimum: 0
        format: "int32"
      responses:
        200:
          description: "search results matching criteria"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/PlaylistItem"
        400:
          description: "bad input parameter"
      x-swagger-router-controller: "swagger_server.controllers.public_controller"
    post:
      tags:
      - "users"
      summary: "crea una lista de reproducción"
      description: "Un usuario crea una lista de reproducción."
      operationId: "add_playlist"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "playlistItem"
        description: "Playlist item to add"
        required: false
        schema:
          $ref: "#/definitions/PlaylistItemNew"
      responses:
        201:
          description: "item created"
          schema:
            $ref: "#/definitions/PlaylistItem"
        400:
          description: "invalid input, object invalid"
        401:
          description: "need authentification"
      x-swagger-router-controller: "swagger_server.controllers.users_controller"
  /playlists/{playlistID}:
    get:
      tags:
      - "public"
      summary: "obtiene una playlist"
      description: "Obtiene los datos de la playlist identificada por playlistID\n"
      operationId: "get_playlist"
      produces:
      - "application/json"
      parameters:
      - name: "playlistID"
        in: "path"
        description: "ID de la playlist"
        required: true
        type: "string"
      responses:
        200:
          description: "success"
          schema:
            $ref: "#/definitions/PlaylistItem"
        400:
          description: "bad input parameter"
        404:
          description: "not found"
      x-swagger-router-controller: "swagger_server.controllers.public_controller"
    put:
      tags:
      - "users"
      summary: "actualiza la información de una playlist"
      description: "Un usuario actualiza la información de una playlist de su propiedad."
      operationId: "update_playlist"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "playlistID"
        in: "path"
        description: "ID de la playlist"
        required: true
        type: "string"
      - in: "body"
        name: "playlistItem"
        description: "Playlist item to update"
        required: false
        schema:
          $ref: "#/definitions/PlaylistItemNew"
      responses:
        200:
          description: "item updated"
          schema:
            $ref: "#/definitions/PlaylistItem"
        400:
          description: "invalid input, object invalid"
        401:
          description: "need authentification"
        403:
          description: "operation forbidden for user"
        404:
          description: "not found"
      x-swagger-router-controller: "swagger_server.controllers.users_controller"
    delete:
      tags:
      - "users"
      summary: "elimina una playlist"
      description: "Elimina la playlist identificada por playlistID\n"
      operationId: "delete_playlist"
      produces:
      - "application/json"
      parameters:
      - name: "playlistID"
        in: "path"
        description: "ID de la playlist"
        required: true
        type: "string"
      responses:
        200:
          description: "playlist deleted"
        400:
          description: "bad input parameter"
        401:
          description: "need authentification"
        403:
          description: "operation forbidden for user"
        404:
          description: "not found"
      x-swagger-router-controller: "swagger_server.controllers.users_controller"
  /playlists{playlistID}/songs:
    post:
      tags:
      - "users"
      summary: "añade una canción a una lista de reproducción"
      description: "Un usuario añade una canción a una lista de reproducción de su\
        \ propiedad."
      operationId: "add_playlist_song"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "playlistID"
        in: "path"
        description: "ID de la playlist"
        required: true
        type: "string"
      - in: "body"
        name: "songID"
        description: "Song to add"
        required: false
        schema:
          type: "string"
          format: "int32"
      responses:
        201:
          description: "song added"
          schema:
            $ref: "#/definitions/PlaylistItem"
        400:
          description: "invalid input, object invalid"
        401:
          description: "need authentification"
        403:
          description: "operation forbidden for user"
        404:
          description: "not found"
      x-swagger-router-controller: "swagger_server.controllers.users_controller"
  /playlists/{playlistID}/songs/{songID}:
    delete:
      tags:
      - "users"
      summary: "elimina una canción de una playlist identificada por playlistID"
      description: "Elimina una canción de la playlist identificada por playlistID\n"
      operationId: "delete_playlist_song"
      produces:
      - "application/json"
      parameters:
      - name: "playlistID"
        in: "path"
        description: "ID de la playlist"
        required: true
        type: "string"
      - name: "songID"
        in: "path"
        description: "ID de la canción"
        required: true
        type: "string"
      responses:
        200:
          description: "song deleted"
        400:
          description: "bad input parameter"
        401:
          description: "need authentification"
        403:
          description: "operation forbidden for user"
        404:
          description: "not found"
      x-swagger-router-controller: "swagger_server.controllers.users_controller"
...
  /profiles/{profileID}/follow:
    post:
      tags:
      - "users"
      summary: "comienza a seguir al usuario identificado por profileID"
      description: "Comienza a seguir al usuario identificado por profileID"
      operationId: "follow_profile"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "profileID"
        in: "path"
        description: "ID del perfil"
        required: true
        type: "string"
      responses:
        200:
          description: "followed"
        400:
          description: "invalid input, object invalid"
        401:
          description: "need authentification"
        404:
          description: "not found"
      x-swagger-router-controller: "swagger_server.controllers.users_controller"
    delete:
      tags:
      - "users"
      summary: "deja de seguir al usuario identificado por profileID"
      description: "Deja de seguir al usuario identificado por profileID\n"
      operationId: "unfollow_profile"
      produces:
      - "application/json"
      parameters:
      - name: "profileID"
        in: "path"
        description: "ID del perfil"
        required: true
        type: "string"
      responses:
        200:
          description: "unfollowed"
        400:
          description: "bad input parameter"
        401:
          description: "need authentification"
        404:
          description: "not found"
      x-swagger-router-controller: "swagger_server.controllers.users_controller"
  /account:
    get:
      tags:
      - "users"
      summary: "devuelve la información de la cuenta del usuario"
      description: "Devuelve la información de la cuenta del usuario.\n"
      operationId: "get_account"
      produces:
      - "application/json"
      parameters: []
      responses:
        200:
          description: "Datos de la cuenta del usuario"
          schema:
            $ref: "#/definitions/AccountItem"
        401:
          description: "need authentification"
      x-swagger-router-controller: "swagger_server.controllers.users_controller"
    put:
      tags:
      - "users"
      summary: "actualiza información de cuenta de usuario"
      description: "Actualiza la información de la cuenta del usuario. El campo \"\
        friends\" de los amigos de un usuario está siempre vacío (valor NULL)"
      operationId: "update_account"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "accountItem"
        description: "Datos a actualizar"
        required: false
        schema:
          $ref: "#/definitions/AccountItemUpdate"
      responses:
        200:
          description: "item updated"
          schema:
            $ref: "#/definitions/AccountItem"
        400:
          description: "invalid input, object invalid"
        401:
          description: "need authentification"
      x-swagger-router-controller: "swagger_server.controllers.users_controller"
    delete:
      tags:
      - "users"
      summary: "borra la cuenta del usuario"
      description: "Elimina definitivamente la cuenta del usuario. No se puede deshacer.\
        \ No funciona con cuentas de administrador\n"
      operationId: "delete_account"
      produces:
      - "application/json"
      parameters: []
      responses:
        200:
          description: "Cuenta de usuario eliminada"
        401:
          description: "need authentification"
      x-swagger-router-controller: "swagger_server.controllers.users_controller"
  /account/credentials:
    put:
      tags:
      - "users"
      summary: "modifica credenciales de acceso de cuenta de usuario"
      description: "Modifica las credenciales de acceso de la cuenta del usuario.\
        \ El campo \"friends\" de los amigos de un usuario está siempre vacío (valor\
        \ NULL)"
      operationId: "update_account_credentials"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "loginItem"
        description: "Datos a actualizar"
        required: false
        schema:
          $ref: "#/definitions/LoginItem"
      responses:
        200:
          description: "item updated"
          schema:
            $ref: "#/definitions/AccountItem"
        400:
          description: "invalid input, object invalid"
        401:
          description: "need authentification"
      x-swagger-router-controller: "swagger_server.controllers.users_controller"
  /account/session:
    get:
      tags:
      - "users"
      summary: "devuelve información de sincronización de canciones"
      description: "Devuelve la lista de reproducción, la canción y el segundo que\
        \ estaba escuchando el usuario previamente\n"
      operationId: "restore_session"
      produces:
      - "application/json"
      parameters: []
      responses:
        200:
          description: "Session data"
          schema:
            $ref: "#/definitions/SessionItem"
        401:
          description: "need authentification"
      x-swagger-router-controller: "swagger_server.controllers.users_controller"
    put:
      tags:
      - "users"
      summary: "guarda información de sincronización de canciones"
      description: "Guarda la lista de reproducción, la canción y el segundo que está\
        \ escuchando el usuario"
      operationId: "save_session"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "sessionItem"
        description: "Datos a sincronizar"
        required: false
        schema:
          $ref: "#/definitions/SessionItem"
      responses:
        200:
          description: "session updated"
          schema:
            $ref: "#/definitions/AccountItem"
        400:
          description: "invalid input, object invalid"
        401:
          description: "need authentification"
      x-swagger-router-controller: "swagger_server.controllers.users_controller"
...
definitions:
...
  SongItem:
    type: "object"
    required:
    - "albumID"
    - "albumName"
    - "authorID"
    - "authorName"
    - "genre"
    - "id"
    - "lenght"
    - "name"
    properties:
      id:
        type: "string"
        format: "int32"
        example: "123456789"
      name:
        type: "string"
        example: "I Will Survive"
      lenght:
        type: "string"
        format: "int32"
        example: "03:05"
      authorID:
        type: "string"
        format: "int32"
        example: "4567"
      authorName:
        type: "string"
        example: "Gloria Gaynor"
      albumID:
        type: "string"
        format: "int32"
        example: "9876"
      albumName:
        type: "string"
        example: "Allenrok"
      genre:
        type: "array"
        items:
          type: "string"
          example: "rock"
    example:
      lenght: "03:05"
      albumName: "Allenrok"
      authorName: "Gloria Gaynor"
      name: "I Will Survive"
      genre:
      - "rock"
      - "rock"
      albumID: "9876"
      id: "123456789"
      authorID: "4567"
  PlaylistItemNew:
    type: "object"
    required:
    - "name"
    properties:
      name:
        type: "string"
        example: "Nombre de la nueva lista"
      description:
        type: "string"
        example: "Descripción de la nueva lista"
    example:
      name: "Nombre de la nueva lista"
      description: "Descripción de la nueva lista"
  PlaylistItem:
    type: "object"
    required:
    - "creationDate"
    - "description"
    - "id"
    - "name"
    - "ownerID"
    - "ownerName"
    - "songs"
    properties:
      id:
        type: "string"
        format: "int32"
        example: "98765"
      name:
        type: "string"
        example: "Best music 1980"
      ownerID:
        type: "string"
        format: "int32"
        example: "11111"
      ownerName:
        type: "string"
        example: "Richard Stallman"
      creationDate:
        type: "string"
        format: "date"
        example: "2018-03-11T00:00:00.000Z"
      description:
        type: "string"
        example: "Relax and enjoy the very best hits of the most awesome decade of\
          \ music"
      songs:
        type: "array"
        items:
          $ref: "#/definitions/SongItem"
    example:
      ownerName: "Richard Stallman"
      songs:
      - lenght: "03:05"
        albumName: "Allenrok"
        authorName: "Gloria Gaynor"
        name: "I Will Survive"
        genre:
        - "rock"
        - "rock"
        albumID: "9876"
        id: "123456789"
        authorID: "4567"
      - lenght: "03:05"
        albumName: "Allenrok"
        authorName: "Gloria Gaynor"
        name: "I Will Survive"
        genre:
        - "rock"
        - "rock"
        albumID: "9876"
        id: "123456789"
        authorID: "4567"
      name: "Best music 1980"
      description: "Relax and enjoy the very best hits of the most awesome decade\
        \ of music"
      id: "98765"
      ownerID: "11111"
      creationDate: "2018-03-11T00:00:00.000Z"
...
  FriendItem:
    type: "object"
    required:
    - "bio"
    - "id"
    - "name"
    - "username"
    properties:
      id:
        type: "string"
        format: "int32"
        example: "11111"
      username:
        type: "string"
        example: "rms"
      name:
        type: "string"
        example: "Coleguita Molón"
      bio:
        type: "string"
        example: "Somos muy amigos"
    example:
      name: "Coleguita Molón"
      bio: "Somos muy amigos"
      id: "11111"
      username: "rms"
...
  AccountItemUpdate:
    type: "object"
    properties:
      username:
        type: "string"
        example: "Nuevo username"
      name:
        type: "string"
        example: "Nuevo nombre"
      bio:
        type: "string"
        example: "Nueva bio"
    example:
      name: "Nuevo nombre"
      bio: "Nueva bio"
      username: "Nuevo username"
  AccountItem:
    type: "object"
    required:
    - "bio"
    - "friends"
    - "id"
    - "mail"
    - "name"
    - "playlists"
    - "username"
    properties:
      id:
        type: "string"
        format: "int32"
        example: "11111"
      username:
        type: "string"
        example: "rms"
      name:
        type: "string"
        example: "Richard Stallman"
      bio:
        type: "string"
        example: "¡Larga vida al Open Source!"
      mail:
        type: "string"
        example: "rms@gnu.org"
      friends:
        type: "array"
        items:
          $ref: "#/definitions/FriendItem"
      playlists:
        type: "array"
        items:
          $ref: "#/definitions/PlaylistItem"
    example:
      mail: "rms@gnu.org"
      name: "Richard Stallman"
      playlists:
      - ownerName: "Richard Stallman"
        songs:
        - lenght: "03:05"
          albumName: "Allenrok"
          authorName: "Gloria Gaynor"
          name: "I Will Survive"
          genre:
          - "rock"
          - "rock"
          albumID: "9876"
          id: "123456789"
          authorID: "4567"
        - lenght: "03:05"
          albumName: "Allenrok"
          authorName: "Gloria Gaynor"
          name: "I Will Survive"
          genre:
          - "rock"
          - "rock"
          albumID: "9876"
          id: "123456789"
          authorID: "4567"
        name: "Best music 1980"
        description: "Relax and enjoy the very best hits of the most awesome decade\
          \ of music"
        id: "98765"
        ownerID: "11111"
        creationDate: "2018-03-11T00:00:00.000Z"
      - ownerName: "Richard Stallman"
        songs:
        - lenght: "03:05"
          albumName: "Allenrok"
          authorName: "Gloria Gaynor"
          name: "I Will Survive"
          genre:
          - "rock"
          - "rock"
          albumID: "9876"
          id: "123456789"
          authorID: "4567"
        - lenght: "03:05"
          albumName: "Allenrok"
          authorName: "Gloria Gaynor"
          name: "I Will Survive"
          genre:
          - "rock"
          - "rock"
          albumID: "9876"
          id: "123456789"
          authorID: "4567"
        name: "Best music 1980"
        description: "Relax and enjoy the very best hits of the most awesome decade\
          \ of music"
        id: "98765"
        ownerID: "11111"
        creationDate: "2018-03-11T00:00:00.000Z"
      bio: "¡Larga vida al Open Source!"
      id: "11111"
      friends:
      - name: "Coleguita Molón"
        bio: "Somos muy amigos"
        id: "11111"
        username: "rms"
      - name: "Coleguita Molón"
        bio: "Somos muy amigos"
        id: "11111"
        username: "rms"
      username: "rms"
  LoginItem:
    type: "object"
    required:
    - "mail"
    - "pass"
    properties:
      mail:
        type: "string"
        example: "rms@gnu.org"
      pass:
        type: "string"
        example: "extremellysecurepassword"
    example:
      mail: "rms@gnu.org"
      pass: "extremellysecurepassword"
...
  SessionItem:
    type: "object"
    required:
    - "playlistID"
    - "second"
    - "songID"
    properties:
      playlistID:
        type: "string"
        format: "int32"
        example: "345"
      songID:
        type: "string"
        format: "int32"
        example: "4975"
      second:
        type: "string"
        format: "int32"
        example: "120"
    example:
      playlistID: "345"
      songID: "4975"
      second: "120"